requirements.txt
amqp==2.5.2
argo-workflows==3.0.2
asgiref==3.2.3
attrs==19.3.0
autopep8==1.5.2
billiard==3.6.2.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
configparser==5.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
Django==3.0.3
django-filter==2.3.0
django-pytest==0.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
entrypoints==0.3
factory-boy==2.12.0
Faker==4.1.0
flake8==3.7.9
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.11.2
humanize==2.0.0
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.7
kubernetes==10.0.1
kubetest==0.8.1
MarkupSafe==1.1.1
mccabe==0.6.1
minio==5.0.10
mock==4.0.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
py-timeparser==0.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.10.0
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
responses==0.10.15
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
websocket-client==0.57.0
zipp==3.0.0
